+++ src/modules/mavlink/mavlink_messages.cpp
@@ -119,6 +119,7 @@ using matrix::Vector3f;
 using matrix::wrap_2pi;
 
 #include "streams/AUTOPILOT_VERSION.hpp"
+#include "streams/ESC_STATUS.hpp"
 #include "streams/EXTENDED_SYS_STATE.hpp"
 #include "streams/FLIGHT_INFORMATION.hpp"
 #include "streams/PROTOCOL_VERSION.hpp"
@@ -5381,92 +5382,6 @@ protected:
 	}
 };
 
-class MavlinkStreamESCStatus : public MavlinkStream
-{
-public:
-	const char *get_name() const override
-	{
-		return MavlinkStreamESCStatus::get_name_static();
-	}
-
-	static constexpr const char *get_name_static()
-	{
-		return "ESC_STATUS";
-	}
-
-	static constexpr uint16_t get_id_static()
-	{
-		return MAVLINK_MSG_ID_ESC_STATUS;
-	}
-
-	uint16_t get_id() override
-	{
-		return get_id_static();
-	}
-
-	static MavlinkStream *new_instance(Mavlink *mavlink)
-	{
-		return new MavlinkStreamESCStatus(mavlink);
-	}
-
-	unsigned get_size() override
-	{
-		static constexpr unsigned size_per_batch = MAVLINK_MSG_ID_ESC_STATUS_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES;
-		return _esc_status_sub.advertised() ? size_per_batch * _number_of_batches : 0;
-	}
-
-private:
-	uORB::Subscription _esc_status_sub{ORB_ID(esc_status)};
-
-	/* do not allow top copying this class */
-	MavlinkStreamESCStatus(MavlinkStreamESCStatus &) = delete;
-	MavlinkStreamESCStatus &operator = (const MavlinkStreamESCStatus &) = delete;
-
-protected:
-
-	uint8_t _number_of_batches{0};
-
-	explicit MavlinkStreamESCStatus(Mavlink *mavlink) : MavlinkStream(mavlink)
-	{}
-
-	bool send(const hrt_abstime t) override
-	{
-
-		esc_status_s esc_status;
-		uint8_t batch_size = MAVLINK_MSG_ESC_STATUS_FIELD_RPM_LEN;
-
-		if (_esc_status_sub.update(&esc_status)) {
-
-			mavlink_esc_status_t msg = {};
-
-			msg.time_usec = esc_status.timestamp;
-
-			// Ceil value of integer division. For 1-4 esc => 1 batch, 5-8 esc => 2 batches etc
-			_number_of_batches = ceilf((float)esc_status.esc_count / batch_size);
-
-			for (int batch_number = 0; batch_number < _number_of_batches; batch_number++) {
-
-				msg.index = batch_number * batch_size;
-
-				for (int esc_index = 0; esc_index < batch_size ; esc_index++) {
-
-					msg.rpm[esc_index] = esc_status.esc[esc_index].esc_rpm;
-					msg.voltage[esc_index] = esc_status.esc[esc_index].esc_voltage;
-					msg.current[esc_index] = esc_status.esc[esc_index].esc_current;
-				}
-
-				mavlink_msg_esc_status_send_struct(_mavlink->get_channel(), &msg);
-
-			}
-
-			return true;
-		}
-
-		return false;
-	}
-};
-
-
 static const StreamListItem streams_list[] = {
 	create_stream_list_item<MavlinkStreamHeartbeat>(),
 	create_stream_list_item<MavlinkStreamStatustext>(),
@@ -5531,7 +5446,9 @@ static const StreamListItem streams_list[] = {
 	create_stream_list_item<MavlinkStreamOrbitStatus>(),
 	create_stream_list_item<MavlinkStreamObstacleDistance>(),
 	create_stream_list_item<MavlinkStreamESCInfo>(),
+#if defined(ESC_STATUS_HPP)
 	create_stream_list_item<MavlinkStreamESCStatus>(),
+#endif // ESC_STATUS_HPP
 #if defined(AUTOPILOT_VERSION_HPP)
 	create_stream_list_item<MavlinkStreamAutopilotVersion>(),
 #endif // AUTOPILOT_VERSION_HPP
